+++ Jeux/CrossServeur/ClasseTDCross.py
@@ -50,7 +50,7 @@ class JeuTortuesDuoCross(JeuTortuesCross):
             else:
                 descip+="{0} : {1}\n".format(dictEmote[i.couleur],i.titre)
         
-        embed.add_field(name="<:otCOINS:873226814527520809> gagnés par chacun des gagnants",value="{0} <:otCOINS:873226814527520809>".format(75+sum(self.mises.values())/2))
+        embed.add_field(name="<:otCOINS:873226814527520809> gagnés par chacun des gagnants",value="{0} <:otCOINS:873226814527520809>".format(75+sum(self.paris.mises.values())/2))
 
         embed=auteur(bot.user,None,None,embed,"olbor")
         embed.set_footer(text="OT!tortuesduocross")
@@ -81,7 +81,7 @@ class JeuTortuesDuoCross(JeuTortuesCross):
         connexionOT,curseurOT=connectSQL("OT","Guild","Guild",None,None)
         for i in self.joueurs:
             if i.equipe==team:
-                gainCoins(i.userid,75+sum(self.mises.values())/2)
+                gainCoins(i.userid,75+sum(self.paris.mises.values())/2)
                 count,state=2,"W"
             else:
                 count,state=-1,"L"
