+++ Jeux/Tortues/Tortues.py
@@ -5,6 +5,7 @@ import discord
 from Core.Fonctions.Embeds import createEmbed, embedAssert, exeErrorExcept
 from Jeux.Tortues.ClassesAutres import Carte
 from Jeux.Tortues.ClasseTortues import JeuTortues
+from Jeux.Paris import Pari
 
 listeCouleurs=("rouge","jaune","bleue","verte","violette")
 emotes=["<:ot1:705766186909958185>","<:ot2:705766186989912154>","<:ot3:705766186930929685>","<:ot4:705766186947706934>","<:ot5:705766186713088042>","<:OTTbleu:860119157491892255>", "<:OTTjaune:860119157688631316>", "<:OTTrouge:860119157495693343>", "<:OTTvert:860119157331853333>", "<:OTTviolet:860119157672247326>"]
@@ -17,7 +18,6 @@ async def startGameTortues(ctx,bot,inGame,gamesTortues):
         assert ctx.author.id not in inGame, "Terminez votre partie en cours avant de lancer ou rejoindre une partie."
         game=JeuTortues(ctx.guild,ctx.author.id)
         game.ids.append(ctx.author.id)
-        game.mises[ctx.author.id]=0
         inGame.append(ctx.author.id)
         message=await ctx.send(embed=createEmbed("Course des tortues","Le jeu se joue de 2 à 5 joueurs.\nAu début de la partie, chaque joueur se voit attribuer une couleur secrète, envoyée en message privé, qui est celle de sa tortue.\nLe but est d'atteindre l'arrivée avant tout le monde, en jouant avec des cartes qui font avancer les tortues.\nLes joueurs jouent chacun leur tour. Les réactions <:ot1:705766186909958185> à <:ot5:705766186713088042> permettent de choisir sa carte.\nSi vous choisissez une carte 'au choix', cliquez ensuite sur la réaction de la tortue que vous voulez déplacer <:OTTbleu:860119157491892255> <:OTTjaune:860119157688631316> <:OTTrouge:860119157495693343> <:OTTvert:860119157331853333> <:OTTviolet:860119157672247326>.\nLes cartes 'dernière tortue' font avancer la dernière tortue.\nEn dehors de la case départ, les tortues s'empilent et avancent en même temps !\nSi plusieurs tortues arrivent en même temps, celle qui est le plus bas gagne !\nBonne chance !\n\nAppuyez sur la réaction <:otVALIDER:772766033996021761> pour rejoindre la partie de Course des tortues et <:otANNULER:811242376625782785> pour annuler votre participation.\n<@{0}> peut lancer la partie en appuyant sur <:otVALIDER:772766033996021761>, sinon elle se lancera automatiquement au bout de 1 minute.".format(ctx.author.id),0xad917b,ctx.invoked_with.lower(),ctx.guild))
         gamesTortues[message.id]=game
@@ -55,6 +55,7 @@ async def startGameTortues(ctx,bot,inGame,gamesTortues):
 
         game.giveCards()
         turn=randint(0,len(game.joueurs)-1)
+        game.paris=Pari(game.ids,"Tortues")
         while game.playing:
             await message.edit(embed=game.embedGame(game.joueurs[turn].user))
             couleur,valeur,carte=await game.play(turn,message,bot)
@@ -67,7 +68,9 @@ async def startGameTortues(ctx,bot,inGame,gamesTortues):
                 await message.clear_reactions()
                 await message.unpin()
                 game.stats(game.getWinner())
-                
+                game.paris.distribParis(game.getWinner().userid)
+
+            game.fermeture()
             game.joueurs[turn].jeu.remove(carte)
             game.joueurs[turn].pioche(game.cartes)
             turn+=1
